community _ th 1. 커뮤니티 메인 화면 jsp 변경 2. jdbc로 커뮤니티 전체 데이터 가져오기 작업

+++ semiproject_uglys/src/main/java/com/veg/kth/community/model/dto/Procedure.java
@@ -8,7 +8,7 @@ import lombok.Data;
 public class Procedure {
 
 	private int procedureNo;
-	private int procedureOrder;
+	private Long procedureOrder;
 	private String procedureComment;
 	private String procedureOriginalFileName;
 	private String procedureRenamedFileName;
